Testing the math_sphe_harmonics() function

+++ modules/math.h
@@ -45,7 +45,7 @@
 	double math_assoc_legendre_poly(const size_t l,
 	                                const size_t m, const double x);
 
-	double math_sphe_harmonics(const int l, const int m,
+	double math_sphe_harmonics(const size_t l, const int m,
 	                           const double theta, const double phi);
 
 	double math_wigner_3j(const int a, const int b, const int c,
